move containers to Chat scene containers.
