+++ frontend/src/scenes/Chat/containers/Chat/styles.module.sass
@@ -1,4 +1,4 @@
-.chat__container
+.chatContainer
   display: flex
   flex-direction: column
   height: 100vh // remove when workspace layout done
