+++ frontend/src/scenes/Chat/containers/ChatBody/styles.module.sass
@@ -1,4 +1,4 @@
-.chat__body
+.chatBody
   padding: 0 15px
   display: flex
   flex-direction: column
