+++ frontend/src/scenes/Chat/containers/ChatFooter/styles.module.sass
@@ -1,4 +1,4 @@
-.chat_footer
+.chatFooter
   padding: 0 15px 15px
   min-height: 100px
