+++ frontend/src/scenes/Chat/containers/ChatHeader/styles.module.sass
@@ -1,5 +1,4 @@
-.ChatHeader
-  &__container
+.chatContainer
     height: 64px
     display: flex
     align-items: center
@@ -28,7 +27,7 @@
         .iconChatType
           font-size: 10px
 
-          &__active
+          &.active
             color: var(--icon-active)
 
       .topic
