+++ frontend/src/scenes/Chat/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ChatContainer from '../../containers/Chat';
+import ChatContainer from './containers/Chat';
 
 const ChatScene = () => (
   <ChatContainer />
